requirements.txt
sentence-transformers 
groq
langchain-community
python-dotenv
streamlit